+++ tests/serde.rs
@@ -5,42 +5,24 @@ use caretta_id::{CarettaId, CarettaIdD, CarettaIdQ, CarettaIdS, CarettaIdT};
 
 use serde::{Deserialize, Serialize};
 
-use serde_test::{Token, assert_tokens};
+use serde_test::{Configure, Token, assert_tokens};
 
 #[test]
-fn nil() {
-    assert_tokens(&CarettaId::NIL, &[Token::Str("0000000")])
+fn nil_readable() {
+    assert_tokens(&CarettaId::NIL.readable(), &[Token::Str("0000000")])
 }
 
 #[test]
-fn max() {
-    assert_tokens(&CarettaId::MAX, &[Token::Str("zzzzzzz")]);
+fn max_readable() {
+    assert_tokens(&CarettaId::MAX.readable(), &[Token::Str("zzzzzzz")]);
 }
 
 #[test]
-fn single_nil() {
-    assert_tokens(&CarettaIdS::NIL, &[Token::Str("000")]);
+fn nil_compact() {
+    assert_tokens(&CarettaId::NIL.compact(), &[Token::U64(0)])
 }
 
 #[test]
-fn double_nil() {
-    assert_tokens(&CarettaIdD::NIL, &[Token::Str("000-000")]);
-}
-
-#[test]
-fn triple_nil() {
-    assert_tokens(&CarettaIdT::NIL, &[Token::Str("000-000-000")]);
-}
-
-#[test]
-fn quadruple_nil() {
-    assert_tokens(&CarettaIdQ::NIL, &[Token::Str("000-000-000-000")]);
-}
-
-#[derive(Serialize, Deserialize, PartialEq, Debug)]
-struct CarettaIdList {
-    s: CarettaIdS,
-    d: CarettaIdD,
-    t: CarettaIdT,
-    q: CarettaIdQ,
+fn max_compact() {
+    assert_tokens(&CarettaId::MAX.compact(), &[Token::U64(0x7FFFFFFFF)]);
 }